% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_twig_build.R
\name{decisions}
\alias{decisions}
\title{Add discounts to a twig Markov object}
\usage{
decisions(...)
}
\arguments{
\item{...}{decision names}

\item{states}{names of the states to expand}

\item{lengths}{the length of each tunnel state}
}
\value{
a twig layer with decision names

a twig layer with tunnels

a twig layer with decision names
}
\description{
Add discounts to a twig Markov object
}
\examples{
discounts(payoffs = c("cost", "effectiveness"), discounts = c(0.5, 0.5))
discounts(payoffs = c("cost", "effectiveness"), discounts = c(0.15, 0.15))
Title

tunnels(states = c("S1", "S2"), lengths = c(3, 5))
Add decisions to a twig

decisions("A", "B", "C")
}
