% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_twig_build_markov.R
\name{is_cycle_dep}
\alias{is_cycle_dep}
\title{Check if model is cycle dependent}
\usage{
is_cycle_dep(twig_env)
}
\arguments{
\item{twig_env}{The twig environment object}
}
\value{
Boolean indicating if model is cycle dependent
}
\description{
Checks if any of the functions used in the model take cycle as an argument
}
