% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_twig_build.R
\name{payoffs}
\alias{payoffs}
\title{Add payoffs to a twig}
\usage{
payoffs(names, discount_rates = NULL)
}
\arguments{
\item{names}{... a named list containing the payoff function names}

\item{discount_rates}{... an optional numeric vector containing discount rates per cycle. If null 0 is used}
}
\value{
a twig layer with payoffs
}
\description{
Add payoffs to a twig
}
\examples{
payoffs(cost = cost_function(state), effectiveness = effective_function(state))
}
