% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_twig_build.R
\name{+.twig_class}
\alias{+.twig_class}
\title{Define a method for the \code{+} operator for \code{twig} objects}
\usage{
\method{+}{twig_class}(twig, layer)
}
\arguments{
\item{twig}{}

\item{layer}{}
}
\value{
twig
}
\description{
Adds layers to the twig object
}
\examples{
twig <- twig() + 
decisions("StandardOfCare", "StrategyA", "StrategyB", "StrategyAB")
}
