% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_prob_functions.R
\name{prob2rate}
\alias{prob2rate}
\title{Convert Probability to Rate}
\usage{
prob2rate(prob)
}
\arguments{
\item{prob}{A numeric value (or vector) representing the probability. Must be
in the interval \verb{[0, 1]}. A probability of exactly 1 returns \code{Inf}.}
}
\value{
A numeric value representing the rate.
}
\description{
This function converts a probability to a rate using the formula \code{-log(1 - prob)}.
}
\examples{
prob <- 0.1
rate <- prob2rate(prob)
print(rate)
}
