% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_prob_functions.R
\name{rate2prob}
\alias{rate2prob}
\title{Convert Rate to Probability}
\usage{
rate2prob(rate)
}
\arguments{
\item{rate}{A numeric value (or vector) representing the rate. Rates must be
non-negative.}
}
\value{
A numeric value representing the probability.
}
\description{
This function converts a rate to a probability using the formula \code{1 - exp(-rate)}.
}
\examples{
rate <- 0.1
prob <- rate2prob(rate)
print(prob)
}
