% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_twig_build.R
\name{states}
\alias{states}
\title{Add Markov states to a twig}
\usage{
states(names, init_probs, max_cycle_in_states = NULL)
}
\arguments{
\item{...}{Markov state names}
}
\value{
a twig layer with Markov state names
}
\description{
Add Markov states to a twig
}
\examples{
states("Healthy", "Sick", "Dead")
}
