% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_twig_build.R
\name{twig}
\alias{twig}
\title{Create a new twig}
\usage{
twig()
}
\arguments{
\item{model_type}{}

\item{n_cycles}{= 50}
}
\value{
a new twig object
}
\description{
Create a new twig
}
\examples{
twig(model_type = "Markov", n_cycles = 40)

}
