% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twig_build.R
\name{twig_build.markov_twig}
\alias{twig_build.markov_twig}
\title{Build a Markov model structure}
\usage{
\method{twig_build}{markov_twig}(twig_env)
}
\arguments{
\item{twig_env}{}
}
\value{
twig structure containing formulae strings
}
\description{
takes in a twig formula structure and returns a twig numerical formula structure
}
\examples{
print("see vignettes(package = 'twig')") 
}
