% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twig_build.R
\name{twig_build}
\alias{twig_build}
\title{Builds a Markov model}
\usage{
twig_build(x, ...)
}
\arguments{
\item{twig_env}{}
}
\value{
a twig object containing the equations for the Markov model
}
\description{
Builds a Markov model
}
\examples{
print("see vignettes(package = 'twig')") 
}
