% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twig_build.R
\name{twig_build.twig_markov}
\alias{twig_build.twig_markov}
\title{Build a Markov model structure}
\usage{
\method{twig_build}{twig_markov}(twig_obj, params = NULL, simplify = FALSE)
}
\arguments{
\item{twig_obj}{}
}
\value{
twig structure containing formulae strings
}
\description{
takes in a twig formula structure and returns a twig numerical formula structure
}
\examples{
print("see vignettes(package = 'twig')") 
}
