% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twig_evaluate.R
\name{twig_evaluate}
\alias{twig_evaluate}
\title{Runs a decision tree or Markov model}
\usage{
twig_evaluate(x, ...)
}
\arguments{
\item{model_num_struc}{a matrix containing the numerical twig object from the twig_parse() function}
}
\value{
twig object
}
\description{
runs the decision tree or markov model and returns the outcomes either traces and summary outcomes
}
\examples{
print("see vignettes: vignettes(package='twig')")
}
