% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twig_evaluate.R
\name{twig_evaluate.twig_decision}
\alias{twig_evaluate.twig_decision}
\title{Runs the decision tree}
\usage{
\method{twig_evaluate}{twig_decision}(model_struc, params = NULL)
}
\arguments{
\item{model_num_struc}{a matrix containing the numerical twig decision object from the twig_parse() function}
}
\value{
twig object
}
\description{
runs the decision model and returns the expected payoff values and summary outcomes
}
\examples{
print("see vignettes: vignettes(package='twig')")
}
