% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twig_evaluate.R
\name{twig_evaluate.twig_markov}
\alias{twig_evaluate.twig_markov}
\title{Runs the markov model}
\usage{
\method{twig_evaluate}{twig_markov}(
  model_struc,
  model_function_name = "my_markov_model",
  print_model_function = FALSE
)
}
\arguments{
\item{model_num_struc}{a matrix containing the numerical twig object from the twig_parse() function}
}
\value{
twig object
}
\description{
runs the markov model and returns the traces and summary outcomes
}
\examples{
print("see vignettes: vignettes(package='twig')")
}
