% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twig_expand_functions.R
\name{twig_expand_functions}
\alias{twig_expand_functions}
\title{Title}
\usage{
twig_expand_functions(twig_obj, fun_names = NULL, excel_file_name = NULL)
}
\arguments{
\item{twig_obj}{}

\item{fun_names}{}

\item{excel_file_name}{}
}
\description{
Title
}
