% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twig_gen_model_function.R
\name{twig_gen_model_function.decision_twig}
\alias{twig_gen_model_function.decision_twig}
\title{Title}
\usage{
\method{twig_gen_model_function}{decision_twig}(
  twig_obj,
  model_function_name = "my_decision_model",
  print_model_function = FALSE,
  return_twig_objture = TRUE
)
}
\arguments{
\item{twig_obj}{}

\item{model_function_name}{}

\item{print_model_function}{}
}
\value{
twig_obj and generates model function code
}
\description{
Title
}
\examples{
print("see vignettes: vignettes(package='twig')")
}
