% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twig_gen_model_function.R
\name{twig_gen_model_function.markov_twig}
\alias{twig_gen_model_function.markov_twig}
\title{Runs the markov model}
\usage{
\method{twig_gen_model_function}{markov_twig}(
  twig_obj,
  model_function_name = "my_markov_model",
  print_model_function = FALSE,
  sparse = FALSE,
  return_twig_objture = TRUE
)
}
\arguments{
\item{twig_obj}{a matrix containing the numerical twig object from the twig_parse() function}
}
\value{
twig_obj and generates model function code
}
\description{
runs the markov model and returns the traces and summary outcomes
}
\examples{
print("see vignettes: vignettes(package='twig')")
}
