% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twig_gen_model_function.R
\name{twig_gen_model_function}
\alias{twig_gen_model_function}
\title{Runs a decision tree or Markov model}
\usage{
twig_gen_model_function(x, ...)
}
\arguments{
\item{model_num_struc}{a matrix containing the numerical twig object from the twig_parse() function}
}
\value{
twig_env and generates model function code
}
\description{
runs the decision tree or markov model and returns the outcomes either traces and summary outcomes
}
\examples{
print("see vignettes: vignettes(package='twig')")
}
