% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twig_gen_model_function.R
\name{twig_gen_model_function.twig_decision}
\alias{twig_gen_model_function.twig_decision}
\title{Title}
\usage{
\method{twig_gen_model_function}{twig_decision}(
  model_struc,
  model_function_name = "my_decision_model",
  print_model_function = FALSE,
  return_model_structure = TRUE
)
}
\arguments{
\item{model_struc}{}

\item{model_function_name}{}

\item{print_model_function}{}
}
\value{
model_struc and generates model function code
}
\description{
Title
}
\examples{
print("see vignettes: vignettes(package='twig')")
}
