% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twig_gen_model_function.R
\name{twig_gen_model_function.twig_markov}
\alias{twig_gen_model_function.twig_markov}
\title{Runs the markov model}
\usage{
\method{twig_gen_model_function}{twig_markov}(
  mytwig,
  model_function_name = "my_markov_model",
  print_model_function = FALSE,
  sparse = FALSE,
  return_model_structure = TRUE
)
}
\arguments{
\item{model_struc}{a matrix containing the numerical twig object from the twig_parse() function}
}
\value{
model_struc and generates model function code
}
\description{
runs the markov model and returns the traces and summary outcomes
}
\examples{
print("see vignettes: vignettes(package='twig')")
}
