% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twig_parse.R
\name{twig_parse.markov_twig}
\alias{twig_parse.markov_twig}
\title{Parses a twig Markov object}
\usage{
\method{twig_parse}{markov_twig}(twig_obj, params = NULL)
}
\arguments{
\item{twig_obj}{a twig representing the model structure}

\item{params}{a parameter list containing the model parameters}
}
\value{
model_num_struc
}
\description{
given twig equations and a list of parameters it evaluates the model equations and returns the numerical values
}
\examples{
print("See Vignettes")
}
