% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twig_parse.R
\name{twig_parse}
\alias{twig_parse}
\title{Parses a twig model formula structure}
\usage{
twig_parse(x, ...)
}
\arguments{
\item{twig_env}{a twig object containing the equations for the Markov model}

\item{params}{a list containing the model parameters}
}
\value{
a nuemerical twig object structure
}
\description{
parses the string formulae and computes the numerical outcomes
}
\examples{
print("See Vignettes")
}
