% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twig_parse.R
\name{twig_parse.twig_decision}
\alias{twig_parse.twig_decision}
\title{Parses a twig Decision Tree object}
\usage{
\method{twig_parse}{twig_decision}(model_struc, params = NULL)
}
\arguments{
\item{model_struc}{a twig representing the model structure}

\item{params}{a parameter list containing the model parameters}
}
\value{
model_num_struc
}
\description{
given twig equations and a list of parameters it evaluates the model equations and returns the numerical values
}
\examples{
print("See Vignettes")
}
