% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_twig_obj.R
\name{decisions}
\alias{decisions}
\title{Add decisions to a twig}
\usage{
decisions(names)
}
\arguments{
\item{names}{decision names, a character vector of decision names.  Tehy don't need to be included in quotes.}
}
\value{
a twig layer with decision names
}
\description{
Add decisions to a twig
}
\examples{
decisions(names = c(A, B, C))
decisions(names = c("A", "B", "C"))

}
