% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_twig_build.R
\name{event}
\alias{event}
\title{Add event mapping}
\usage{
event(name, options, probabilities, transitions)
}
\arguments{
\item{name}{}

\item{options}{}

\item{probabilities}{}

\item{transitions}{}
}
\value{
twig layer
}
\description{
Add event mapping
}
\examples{
event_mapping(name = "event_progress", 
options = c(TRUE,FALSE), 
probabilities = c(p_progress_function(state), Inf), 
transitions = c("Severe","curr_state")

}
