% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_twig_build.R
\name{payoffs}
\alias{payoffs}
\title{Add final_outcomes from a decision tree to a twig}
\usage{
payoffs(names, discount_rates = NULL)
}
\arguments{
\item{...}{a named list containing the payoffs and the associated payoff functions}
}
\value{
a twig layer with Decision final_outcome names

a twig layer with payoffs
}
\description{
Add final_outcomes from a decision tree to a twig
}
\examples{
final_outcomes("Alive", "Dead")
Add payoffs to a twig

payoffs(cost = cost_function(state), effectiveness = effective_function(state))
}
