% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_twig_build.R
\name{+.twig_class}
\alias{+.twig_class}
\title{Define a method for the \code{+} operator for \code{twig} objects}
\usage{
\method{+}{twig_class}(twig_obj, layer)
}
\arguments{
\item{twig_obj}{}

\item{layer}{}
}
\value{
twig_obj
}
\description{
Adds layers to the twig object
}
\examples{
twig_obj <- twig(model_type = "Markov", n_cycles = 75) + 
decisions("StandardOfCare", "StrategyA", "StrategyB", "StrategyAB")
}
