% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_twig_obj.R
\name{+.twig_class}
\alias{+.twig_class}
\title{Define a method for the \code{+} operator for \code{twig} objects}
\usage{
\method{+}{twig_class}(twig_obj, layer)
}
\arguments{
\item{twig_obj}{A twig object created by the \code{twig} function.}

\item{layer}{A layer to be added to the twig object.}
}
\value{
The modified twig object with the new layer added.
}
\description{
Adds layers to the twig object. If a states layer is added, the twig object is treated as a Markov model.
}
\details{
This method allows layers to be added to a twig object using the \code{+} operator.
}
\examples{
twig_obj <- twig() + 
  decisions("StandardOfCare", "StrategyA", "StrategyB", "StrategyAB")
}
