% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_twig_build.R
\name{run_twig}
\alias{run_twig}
\title{Add payoffs to a twig}
\usage{
run_twig(
  twig,
  params = NULL,
  n_cycles = NULL,
  return_prob = FALSE,
  return_trace = FALSE,
  return_total_payoff = FALSE,
  check_prob_add_to_one = FALSE,
  return_mean_payoff = TRUE
)
}
\arguments{
\item{twig}{... a twig object generated with twig()}

\item{params}{... a list or data.table containing the PSA model parameters (rows=samples, columns=variables)}

\item{n_cycles}{... number of cycles}

\item{return_prob}{... whether to return the probabilities}

\item{return_trace}{... whether to return the trace}

\item{return_total_payoff}{... whether to return total payoffs}

\item{check_prob_add_to_one}{... whether to check the results of the probabilities to make sure they add to 1}

\item{return_mean_payoff}{... wheahter to return the mean of payoffs over all PSA simulations}
}
\value{
a list of results
}
\description{
Add payoffs to a twig
}
\examples{
run_twig(twig=twig, params=params, n_cycles=10)
}
