% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_twig_build.R
\name{states}
\alias{states}
\title{Add Markov states to a twig}
\usage{
states(names, init_probs, tunnel_lengths = NULL)
}
\arguments{
\item{names}{... Markov state names}

\item{init_probs}{... initial probabilities}

\item{tunnel_lengths}{... optional max tunnel lenghts. If ignored a length of 1 is assumed.}
}
\value{
a twig layer with Markov state names
}
\description{
Add Markov states to a twig
}
\examples{
states("Healthy", "Sick", "Dead")
}
