% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_twig_obj.R
\name{twig}
\alias{twig}
\title{Create a new twig object}
\usage{
twig()
}
\value{
A new twig object of class \code{decision_twig} and \code{twig_class} by default.
}
\description{
This function initializes a new twig object, which can be used to build Markov models and decision trees.
}
\examples{
twig_obj <- twig()
# see vignettes for more
}
