% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twig_build.R
\name{twig_build.decision_twig}
\alias{twig_build.decision_twig}
\title{Build a decision tree structure}
\usage{
\method{twig_build}{decision_twig}(twig_obj)
}
\arguments{
\item{twig_obj}{}
}
\value{
twig structure containing formulae strings
}
\description{
takes in a twig model syntax and returns a twig formula structure
}
\examples{
print("see vignettes(package = 'twig')") 
}
